src/backup-job/pyproject.toml
[project]
name = "apic-vibe-portal-backup-job"
version = "0.0.0"
description = "Azure Container Job — Azure API Center backup pipeline for APIC Vibe Portal"
requires-python = ">=3.14"
dependencies = [
    "apic-client",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.21.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "azure-storage-blob>=12.23.0",
    "opentelemetry-api>=1.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.13.1",
    "structlog>=25.5.0",
]

[tool.uv.sources]
apic-client = { path = "../apic_client", editable = true }

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apic_backup"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::SyntaxWarning",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
